Revert "feat: notify on outgoing request changes"
