+++ src/erp.mgt.mn/components/OutgoingRequestWidget.jsx
@@ -1,30 +1,56 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import useOutgoingRequestCount from '../hooks/useOutgoingRequestCount.js';
 
 export default function OutgoingRequestWidget() {
   const navigate = useNavigate();
-  const { counts, hasNew } = useOutgoingRequestCount();
+  const [counts, setCounts] = useState({ pending: 0, accepted: 0, declined: 0 });
 
-  const badgeStyle = {
-    display: 'inline-block',
-    backgroundColor: 'red',
-    color: 'white',
-    borderRadius: '50%',
-    padding: '0.25rem 0.5rem',
-    minWidth: '1.5rem',
-    textAlign: 'center',
-    marginLeft: '0.5rem',
-  };
+  useEffect(() => {
+    let cancelled = false;
+    const statuses = ['pending', 'accepted', 'declined'];
 
-  const anyNew = hasNew.pending || hasNew.accepted || hasNew.declined;
+    async function fetchCounts() {
+      try {
+        const results = await Promise.all(
+          statuses.map(async (status) => {
+            const params = new URLSearchParams({ status });
+            const res = await fetch(
+              `/api/pending_request/outgoing?${params.toString()}`,
+              {
+                credentials: 'include',
+                skipLoader: true,
+              },
+            );
+            if (!res.ok) return 0;
+            const data = await res.json().catch(() => 0);
+            if (typeof data === 'number') return data;
+            if (Array.isArray(data)) return data.length;
+            return Number(data?.count) || 0;
+          }),
+        );
+        if (!cancelled) {
+          setCounts({
+            pending: results[0],
+            accepted: results[1],
+            declined: results[2],
+          });
+        }
+      } catch {
+        if (!cancelled) {
+          setCounts({ pending: 0, accepted: 0, declined: 0 });
+        }
+      }
+    }
+
+    fetchCounts();
+    return () => {
+      cancelled = true;
+    };
+  }, []);
 
   return (
     <div>
-      <h3>
-        Outgoing requests
-        {anyNew && <span style={badgeStyle}>!</span>}
-      </h3>
+      <h3>Outgoing requests</h3>
       <div style={{ display: 'flex', gap: '1rem' }}>
         <div style={{ flex: 1 }}>
           <div style={{ fontSize: '0.9rem', color: '#555' }}>Pending</div>
